+++ app/Config/Validation.php
@@ -24,7 +24,8 @@ class Validation
 		FormatRules::class,
 		FileRules::class,
 		CreditCardRules::class,
-		\Myth\Auth\Authentication\Passwords\ValidationRules::class
+		\Myth\Auth\Authentication\Passwords\ValidationRules::class,
+		\App\Validation\UserUpdateRules::class
 	];
 
 	/**